move code to recorder vm
